# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Egypt Advocates law firm API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/invoices:
    get:
      operationId: listAdminInvoices
      tags: [admin-invoices]
      parameters:
        - in: query
          name: status
          required: false
          schema: { type: string }
        - in: query
          name: clientId
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      operationId: createAdminInvoice
      tags: [admin-invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvoiceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /admin/invoices/{id}:
    get:
      operationId: getAdminInvoice
      tags: [admin-invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateAdminInvoice
      tags: [admin-invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateInvoiceInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /admin/invoices/{id}/mark-paid:
    post:
      operationId: markInvoicePaid
      tags: [admin-invoices]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarkInvoicePaidInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }

  # ---------------- Admin Payments ----------------
  /admin/payments:
    get:
      operationId: listAdminPayments
      tags: [admin-payments]
      parameters:
        - in: query
          name: status
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Payment" }
  /admin/payments/{id}/confirm:
    post:
      operationId: confirmPayment
      tags: [admin-payments]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Payment" }

  # ---------------- Admin Chat ----------------
...
  /admin/legal-articles:
    get:
      operationId: listAdminLegalArticles
      tags: [admin-content]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LegalArticle" }
    post:
      operationId: createAdminLegalArticle
      tags: [admin-content]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLegalArticleInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalArticle" }
  /admin/legal-articles/{id}:
    patch:
      operationId: updateAdminLegalArticle
      tags: [admin-content]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateLegalArticleInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalArticle" }
    delete:
      operationId: deleteAdminLegalArticle
      tags: [admin-content]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }

  # ---------------- Admin Content (Blog Posts) ----------------
...
  /admin/services:
    get:
      operationId: listAdminServices
      tags: [admin-content]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Service" }
    post:
      operationId: createAdminService
      tags: [admin-content]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateServiceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
  /admin/services/{id}:
    patch:
      operationId: updateAdminService
      tags: [admin-content]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateServiceInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
    delete:
      operationId: deleteAdminService
      tags: [admin-content]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /admin/lawyers:
    get:
      operationId: listAdminLawyers
      tags: [admin-content]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lawyer" }
    post:
      operationId: createAdminLawyer
      tags: [admin-content]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLawyerInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lawyer" }
  /admin/lawyers/{id}:
    patch:
      operationId: updateAdminLawyer
      tags: [admin-content]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateLawyerInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lawyer" }
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    Lawyer:
      type: object
      required: [id, slug, nameAr, nameEn, titleAr, titleEn, bioAr, bioEn, photoUrl, email, phone, specializations, yearsExperience, sortOrder, isActive]
      properties:
        id: { type: integer }
        slug: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        titleAr: { type: string }
        titleEn: { type: string }
        bioAr: { type: string }
        bioEn: { type: string }
        photoUrl: { type: ["string", "null"] }
        email: { type: string }
        phone: { type: string }
        specializations:
          type: array
          items: { type: string }
        yearsExperience: { type: integer }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    CreateLawyerInput:
      type: object
      required: [slug, nameAr, nameEn, titleAr, titleEn, bioAr, bioEn, email, phone, specializations, yearsExperience]
      properties:
        slug: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        titleAr: { type: string }
        titleEn: { type: string }
        bioAr: { type: string }
        bioEn: { type: string }
        photoUrl: { type: ["string", "null"] }
        email: { type: string }
        phone: { type: string }
        specializations:
          type: array
          items: { type: string }
        yearsExperience: { type: integer }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    UpdateLawyerInput:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        titleAr: { type: string }
        titleEn: { type: string }
        bioAr: { type: string }
        bioEn: { type: string }
        photoUrl: { type: ["string", "null"] }
        email: { type: string }
        phone: { type: string }
        specializations:
          type: array
          items: { type: string }
        yearsExperience: { type: integer }
        sortOrder: { type: integer }
        isActive: { type: boolean }
    Service:
      type: object
      required: [id, slug, nameAr, nameEn, descriptionAr, descriptionEn, durationMinutes, priceEgp, deliveryMode, isActive]
      properties:
        id: { type: integer }
        slug: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        durationMinutes: { type: integer }
        priceEgp: { type: number }
        deliveryMode: { type: string, enum: [online, in_office, both] }
        practiceAreaId: { type: ["integer", "null"] }
        isActive: { type: boolean }
    CreateServiceInput:
      type: object
      required: [slug, nameAr, nameEn, descriptionAr, descriptionEn, durationMinutes, priceEgp, deliveryMode]
      properties:
        slug: { type: string }
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        durationMinutes: { type: integer }
        priceEgp: { type: number }
        deliveryMode: { type: string, enum: [online, in_office, both] }
        practiceAreaId: { type: ["integer", "null"] }
        isActive: { type: boolean }
    UpdateServiceInput:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        durationMinutes: { type: integer }
        priceEgp: { type: number }
        deliveryMode: { type: string, enum: [online, in_office, both] }
        practiceAreaId: { type: ["integer", "null"] }
        isActive: { type: boolean }
...
    LegalArticle:
      type: object
      required: [id, slug, categoryId, titleAr, titleEn, summaryAr, summaryEn, contentAr, contentEn, lawNumber, year, tags, publishedAt, isPublished, createdAt]
      properties:
        id: { type: integer }
        slug: { type: string }
        categoryId: { type: integer }
        categoryNameAr: { type: ["string", "null"] }
        categoryNameEn: { type: ["string", "null"] }
        titleAr: { type: string }
        titleEn: { type: string }
        summaryAr: { type: string }
        summaryEn: { type: string }
        contentAr: { type: string }
        contentEn: { type: string }
        lawNumber: { type: ["string", "null"] }
        year: { type: ["integer", "null"] }
        tags:
          type: array
          items: { type: string }
        publishedAt: { type: ["string", "null"], format: date-time }
        isPublished: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    CreateLegalArticleInput:
      type: object
      required: [slug, categoryId, titleAr, titleEn, summaryAr, summaryEn, contentAr, contentEn]
      properties:
        slug: { type: string }
        categoryId: { type: integer }
        titleAr: { type: string }
        titleEn: { type: string }
        summaryAr: { type: string }
        summaryEn: { type: string }
        contentAr: { type: string }
        contentEn: { type: string }
        lawNumber: { type: ["string", "null"] }
        year: { type: ["integer", "null"] }
        tags:
          type: array
          items: { type: string }
        isPublished: { type: boolean }
    UpdateLegalArticleInput:
      type: object
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        summaryAr: { type: string }
        summaryEn: { type: string }
        contentAr: { type: string }
        contentEn: { type: string }
        lawNumber: { type: ["string", "null"] }
        year: { type: ["integer", "null"] }
        tags:
          type: array
          items: { type: string }
        isPublished: { type: boolean }
...
    InvoiceItem:
      type: object
      required: [description, quantity, unitPriceEgp]
      properties:
        description: { type: string }
        quantity: { type: number }
        unitPriceEgp: { type: number }
    Invoice:
      type: object
      required: [id, invoiceNumber, clientId, clientName, items, subtotal, tax, total, currency, status, issueDate, createdAt]
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        clientId: { type: integer }
        clientName: { type: string }
        caseId: { type: ["integer", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        currency: { type: string }
        status: { type: string, enum: [draft, sent, paid, overdue, cancelled] }
        dueDate: { type: ["string", "null"], format: date }
        issueDate: { type: string, format: date }
        paidAt: { type: ["string", "null"], format: date-time }
        paymentMethod: { type: ["string", "null"], enum: [instapay, vodafone_cash, fawry, visa, cash, bank_transfer, null] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CreateInvoiceInput:
      type: object
      required: [clientId, items, status, issueDate]
      properties:
        clientId: { type: integer }
        caseId: { type: ["integer", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        taxPercentage: { type: number }
        status: { type: string, enum: [draft, sent, paid, overdue, cancelled] }
        dueDate: { type: ["string", "null"], format: date }
        issueDate: { type: string, format: date }
        notes: { type: ["string", "null"] }
    UpdateInvoiceInput:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/InvoiceItem" }
        taxPercentage: { type: number }
        status: { type: string, enum: [draft, sent, paid, overdue, cancelled] }
        dueDate: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
    MarkInvoicePaidInput:
      type: object
      required: [paymentMethod]
      properties:
        paymentMethod: { type: string, enum: [instapay, vodafone_cash, fawry, visa, cash, bank_transfer] }
        referenceNumber: { type: ["string", "null"] }
        paidAt: { type: ["string", "null"], format: date-time }
    Payment:
      type: object
      required: [id, amountEgp, method, status, createdAt]
      properties:
        id: { type: integer }
        invoiceId: { type: ["integer", "null"] }
        invoiceNumber: { type: ["string", "null"] }
        appointmentId: { type: ["integer", "null"] }
        clientName: { type: ["string", "null"] }
        amountEgp: { type: number }
        method: { type: string, enum: [instapay, vodafone_cash, fawry, visa, cash, bank_transfer] }
        status: { type: string, enum: [pending, confirmed, failed, refunded] }
        referenceNumber: { type: ["string", "null"] }
        paidAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }